// frontend/src/types/chain/hero.ts
import { ethers } from 'ethers'

export interface IChainHero {
  black: ethers.BigNumberish
  earth: ethers.BigNumberish
  energy: ethers.BigNumberish
  exp: ethers.BigNumberish
  fire: ethers.BigNumberish
  level: ethers.BigNumberish
  name: string
  water: ethers.BigNumberish
  white: ethers.BigNumberish
}